arreglando conexion controladores fxml
