--- src/main/java/es/iesfranciscodelosrios/red/DAO/CommentDao.java
@@ -1,4 +1,0 @@
-package es.iesfranciscodelosrios.red.DAO;
-
-public class CommentDao {
-}
