--- src/main/java/es/iesfranciscodelosrios/red/Start.java
@@ -1,7 +1,0 @@
-package es.iesfranciscodelosrios.red;
-
-public class Start {
-    public static void main(String[] args) {
-        App.main(args);
-    }
-}
