--- src/main/java/es/iesfranciscodelosrios/red/controllers/CreatePostC.java
@@ -1,4 +1,0 @@
-package es.iesfranciscodelosrios.red.controllers;
-
-public class CreatePostC {
-}
