--- src/main/java/es/iesfranciscodelosrios/red/controllers/PostC.java
@@ -1,4 +1,0 @@
-package es.iesfranciscodelosrios.red.controllers;
-
-public class PostC {
-}
