--- src/main/java/es/iesfranciscodelosrios/red/interfaces/Dao.java
@@ -1,11 +1,0 @@
-package es.iesfranciscodelosrios.red.interfaces;
-
-public interface Dao {
-    public void save();
-    public void delete();
-    public void update();
-
-    private Object getById(Object o) {
-        return o;
-    }
-}
