+++ src/main/java/redSocial/DAO/CommentDao.java
@@ -1,0 +1,4 @@
+package redSocial.DAO;
+
+public class CommentDao {
+}
