+++ src/main/java/redSocial/Start.java
@@ -1,0 +1,7 @@
+package redSocial;
+
+public class Start {
+    public static void main(String[] args) {
+        App.main(args);
+    }
+}
