+++ src/main/java/redSocial/controllers/CreatePostC.java
@@ -1,0 +1,4 @@
+package redSocial.controllers;
+
+public class CreatePostC {
+}
