+++ src/main/java/redSocial/controllers/HomeC.java
@@ -1,0 +1,4 @@
+package redSocial.controllers;
+
+public class HomeC {
+}
