+++ src/main/java/redSocial/controllers/PostC.java
@@ -1,0 +1,4 @@
+package redSocial.controllers;
+
+public class PostC {
+}
