--- src/main/resources/es/iesfranciscodelosrios/red/fxml/LogIn.fxml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="es.iesfranciscodelosrios.red.controllers.LogInC">
-   <children>
-      <ImageView fitHeight="68.0" fitWidth="93.0" layoutX="254.0" layoutY="31.0" pickOnBounds="true" preserveRatio="true" />
-      <Button fx:id="Login" layoutX="124.0" layoutY="272.0" mnemonicParsing="false" onAction="#LogIn" prefHeight="36.0" prefWidth="341.0" style="-fx-background-color: white; -fx-border-color: #6BA0E5;" text="Log In" />
-      <Button fx:id="Signup" layoutX="123.0" layoutY="319.0" mnemonicParsing="false" onAction="#switchToRegister" prefHeight="36.0" prefWidth="342.0" style="-fx-background-color: #6BA0E5; -fx-text-fill: white;" text="Sign Up" />
-      <PasswordField fx:id="Password" layoutX="123.0" layoutY="210.0" prefHeight="36.0" prefWidth="342.0" promptText="Enter your password" />
-      <TextField fx:id="Nickname" layoutX="123.0" layoutY="138.0" prefHeight="36.0" prefWidth="342.0" promptText="Enter your nickname" />
-      <Label layoutX="126.0" layoutY="109.0" text="Username" />
-      <Label layoutX="126.0" layoutY="182.0" text="Password" />
-   </children>
-</AnchorPane>
