+++ src/main/resources/redSocial.controllers/SignUp.fxml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.PasswordField?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane fx:id="anchorPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="redSocial.controllers.SignUpC">
+    <children>
+        <ImageView fitHeight="77.0" fitWidth="104.0" layoutX="242.0" layoutY="14.0" pickOnBounds="true" preserveRatio="true" />
+        <Button fx:id="SignUp" layoutX="123.0" layoutY="317.0" mnemonicParsing="false" onAction="#SignUp" prefHeight="36.0" prefWidth="342.0" style="-fx-background-color: #6BA0E5; -fx-text-fill: white;" text="Sign Up" />
+        <PasswordField fx:id="Password" layoutX="129.0" layoutY="182.0" prefHeight="36.0" prefWidth="342.0" promptText="Enter your password" />
+        <Label layoutX="132.0" layoutY="155.0" text="Password" />
+      <PasswordField fx:id="Password2" layoutX="129.0" layoutY="258.0" prefHeight="36.0" prefWidth="342.0" promptText="Confirm your Password" />
+      <Label layoutX="129.0" layoutY="227.0" text="Confirm Password" />
+      <TextField fx:id="Username" layoutX="129.0" layoutY="111.0" prefHeight="36.0" prefWidth="342.0" promptText="Enter your usermail" />
+      <Label layoutX="130.0" layoutY="89.0" text="Username" />
+      <Button fx:id="LogIn" layoutX="13.0" layoutY="14.0" mnemonicParsing="false" onAction="#switchToLogIn" prefHeight="36.0" prefWidth="82.0" style="-fx-background-color: #6BA0E5; -fx-text-fill: white;" text="volver" />
+    </children>
+
+</AnchorPane>
